requirements.txt
requests
geopandas
numpy
shapely
scipy
folium
pandas
google-cloud-bigquery
google-cloud-storage
aiohttp